refactor(web): move auth code into feature folder
